chore: rename the drawable leaflet map
