+++ src/lib/maps/drawing/index.js
@@ -1,0 +1,5 @@
+import MapDraw from './mapdraw'
+
+export {
+  MapDraw
+}
