--- src/lib/maps/map-draw/index.js
@@ -1,5 +1,0 @@
-import MapDraw from './map-draw'
-
-export {
-  MapDraw
-}
